import item presets

+++ backend/migrations/2024-12-17-120244_initial_setup/up.sql
@@ -9,7 +9,7 @@ CREATE TABLE item_preset (
     uuid TEXT NOT NULL PRIMARY KEY,
     name TEXT NOT NULL,
     price INTEGER NOT NULL,
-    weight INTEGER NOT NULL,
+    weight REAL NOT NULL,
     description TEXT NOT NULL,
     creator TEXT NOT NULL,
     item_type TEXT NOT Null
